ExactCalculator: Add haptic image button widget
This is a basic extension of ImageButton that provides haptic feedback when
the user touches the button (*not* when they release, to reduce latency).
It uses the system's predefined virtual key haptic feedback effect for
modern vibration effects that the user is already familiar with and
automatic response to changes in system vibration settings.

Change-Id: Ib390860809438b9289364c8a2ece2528853bb338
Signed-off-by: Mohammad Hasan Keramat J <ikeramat@protonmail.com>